+++ app/models/Encashment.js
@@ -1,0 +1,12 @@
+class Encashment {
+
+    constructor(id, timestamp, prevEncashmentId, createdAt) {
+        this.id = id
+        this.timestamp = timestamp
+        this.prevEncashmentId = prevEncashmentId
+        this.createdAt = createdAt
+    }
+
+}
+
+module.exports = Encashment