+++ app/models/Receipt.js
@@ -1,0 +1,10 @@
+class Receipt {
+
+    constructor(timestamp, status) {
+        this.timestamp = timestamp
+        this.status = status
+    }
+
+}
+
+module.exports = Receipt