+++ app/models/dto/ReceiptDTO.js
@@ -1,0 +1,9 @@
+class ReceiptDTO {
+
+    constructor({status, timestamp}) {
+        this.timestamp = timestamp
+        this.status = status
+    }
+}
+
+module.exports = ReceiptDTO