+++ app/resolvers/default/index.js
@@ -5,6 +5,8 @@ const MachineResolver = require("./MachineResolver")
 const ItemMatrixResolver = require("./ItemMatrixResolver")
 const ItemResolver = require("./ItemResolver")
 const DepositResolver = require("./DepositResolver")
+const EncashmentResolver = require("./EncashmentResolver")
+const SaleResolver = require("./SaleResolver")
 
 function DefaultResolvers({controllerService, saleService, notificationSettingsService, billingService, serviceService, itemService, machineService, itemMatrixService, kktService, userService}) {
     const controllerResolver = new ControllerResolver({controllerService, saleService, serviceService, machineService})
@@ -14,6 +16,8 @@ function DefaultResolvers({controllerService, saleService, notificationSettingsS
     const itemMatrixResolver = new ItemMatrixResolver({itemMatrixService})
     const itemResolver = new ItemResolver({saleService})
     const depositResolver = new DepositResolver({billingService})
+    const encashmentResolver = new EncashmentResolver({machineService})
+    const saleResolver = new SaleResolver({saleService})
 
     return {
         Controller: controllerResolver,
@@ -22,7 +26,9 @@ function DefaultResolvers({controllerService, saleService, notificationSettingsS
         Machine: machineResolver,
         ItemMatrix: itemMatrixResolver,
         Item: itemResolver,
-        Deposit: depositResolver
+        Deposit: depositResolver,
+        Encashment: encashmentResolver,
+        Sale: saleResolver
     }
 }