+++ app/services/MachineService.js
@@ -5,6 +5,7 @@ const MachineTypeNotFound = require("../errors/MachineTypeNotFound")
 const ControllerNotFound = require("../errors/ControllerNotFound")
 const EquipmentNotFound = require("../errors/EquipmentNotFound")
 const Machine = require("../models/Machine")
+const Encashment = require("../models/Encashment")
 const MachineLog = require("../models/MachineLog")
 const MachineGroup = require("../models/MachineGroup")
 const MachineType = require("../models/MachineType")
@@ -12,8 +13,9 @@ const Permission = require("../enum/Permission")
 
 class MachineService {
 
-    constructor({MachineModel, MachineGroupModel, MachineTypeModel, MachineLogModel, equipmentService, itemMatrixService, controllerService}) {
+    constructor({MachineModel, EncashmentModel, MachineGroupModel, MachineTypeModel, MachineLogModel, equipmentService, itemMatrixService, controllerService}) {
         this.Machine = MachineModel
+        this.Encashment = EncashmentModel
         this.MachineGroup = MachineGroupModel
         this.MachineType = MachineTypeModel
         this.MachineLog = MachineLogModel
@@ -34,6 +36,10 @@ class MachineService {
         this.getMachineTypeById = this.getMachineTypeById.bind(this)
         this.getAllMachineTypes = this.getAllMachineTypes.bind(this)
         this.addLog = this.addLog.bind(this)
+        this.createEncashment = this.createEncashment.bind(this)
+        this.getEncashmentById = this.getEncashmentById.bind(this)
+        this.getMachineEncashments = this.getMachineEncashments.bind(this)
+        this.getLastMachineEncashment = this.getLastMachineEncashment.bind(this)
     }
 
     async createMachine(input, user) {
@@ -51,7 +57,7 @@ class MachineService {
             machine.place = place
             machine.controller_id = controllerId
             machine.user_id = user.id
-            if(kktId){
+            if (kktId) {
                 machine.kktId = kktId
             }
 
@@ -143,7 +149,6 @@ class MachineService {
             machine.place = place
         }
         machine.kktId = kktId
-
 
 
         return await machine.save()
@@ -297,6 +302,66 @@ class MachineService {
         machineLog.type = type
 
         return await this.MachineLog.create(machineLog, {transaction})
+    }
+
+
+    async createEncashment(machineId, timestamp, user) {
+        if (!user || !user.checkPermission(Permission.CREATE_ENCASHMENT)) {
+            throw new NotAuthorized()
+        }
+
+        const prevEncashment = await this.getLastMachineEncashment(machineId, user)
+
+        const encashment = new Encashment()
+        encashment.timestamp = timestamp
+        encashment.createdAt = new Date()
+        encashment.machine_id = machineId
+
+        if (prevEncashment) {
+            encashment.prevEncashmentId = prevEncashment.id
+        }
+
+        return await this.Encashment.create(encashment)
+    }
+
+
+    async getEncashmentById(encashmentId, user) {
+        if (!user || !user.checkPermission(Permission.GET_MACHINE_ENCASHMENTS)) {
+            throw new NotAuthorized()
+        }
+
+        return await this.Encashment.findOne({
+            where: {
+                id: encashmentId
+            }
+        })
+    }
+
+    async getLastMachineEncashment(machineId, user) {
+        if (!user || !user.checkPermission(Permission.GET_MACHINE_ENCASHMENTS)) {
+            throw new NotAuthorized()
+        }
+
+        return await this.Encashment.findOne({
+            where: {
+                machine_id: machineId
+            },
+            order: [
+                ["id", "DESC"],
+            ]
+        })
+    }
+
+    async getMachineEncashments(machineId, user) {
+        if (!user || !user.checkPermission(Permission.GET_MACHINE_ENCASHMENTS)) {
+            throw new NotAuthorized()
+        }
+
+        return await this.Encashment.findAll({
+            where: {
+                machine_id: machineId
+            }
+        })
     }
 }